client/client_software/requirements.txt
colorzero==2.0
gpiozero==2.0.1
paho-mqtt==2.1.0
setuptools==80.9.0
